+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: dvan-kle <dvan-kle@student.42.fr>            +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/05 14:29:49 by dvan-kle      #+#    #+#                 */
-/*   Updated: 2023/04/20 21:53:53 by dvan-kle      ########   odam.nl         */
+/*   Updated: 2023/05/20 15:57:54 by dvan-kle      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,39 +16,39 @@
 # include <stdlib.h>
 # include <unistd.h>
 
-int		ft_atoi(const char *str);
-void	*ft_bzero(void	*b, size_t n);
-void	*ft_calloc(size_t count, size_t size);
-int		ft_isalnum(int c);
-int		ft_isalpha(int c);
-int		ft_isascii(int c);
-int		ft_isdigit(int c);
-int		ft_isprint(int c);
-char	*ft_itoa(int n);
-void	*ft_memchr(const void *s, int c, size_t n);
-int		ft_memcmp(const void *s1, const void *s2, size_t n);
-void	*ft_memcpy(void *dest, const void *src, size_t n);
-void	*ft_memset(void	*b, int c, size_t n);
-void	*ft_memmove(void *dst, const void *sc, size_t len);
-void	ft_putchar_fd(char c, int fd);
-void	ft_putstr_fd(char *str, int fd);
-void	ft_putendl_fd(char *str, int fd);
-void	ft_putnbr_fd(int nb, int fd);
-char	*ft_strchr(const char *str, int c);
-char	*ft_strdup(const char *src);
-char	*ft_strjoin(char const *s1, char const *s2);
-size_t	ft_strlcat(char *dest, const char *src, size_t size);
-size_t	ft_strlcpy(char *dest, const char *src, size_t size);
-size_t	ft_strlen(const char *str);
-int		ft_strncmp(const char *s1, const char *s2, size_t n);
-char	*ft_strnstr(const char *str, const char *find, size_t len);
-char	*ft_strrchr(const char *str, int c);
-char	*ft_strtrim(char const *s1, char const *set);
-char	*ft_substr(char const *str, unsigned int start, size_t len);
-char	*ft_strmapi(char const *str, char (*f)(unsigned int, char));
-void	ft_striteri(char *str, void (*f)(unsigned int, char*));
-char	**ft_split(char const *str, char c);
-int		ft_tolower(int c);
-int		ft_toupper(int c);
+long long	ft_atoi(const char *str);
+void		*ft_bzero(void	*b, size_t n);
+void		*ft_calloc(size_t count, size_t size);
+int			ft_isalnum(int c);
+int			ft_isalpha(int c);
+int			ft_isascii(int c);
+int			ft_isdigit(int c);
+int			ft_isprint(int c);
+char		*ft_itoa(int n);
+void		*ft_memchr(const void *s, int c, size_t n);
+int			ft_memcmp(const void *s1, const void *s2, size_t n);
+void		*ft_memcpy(void *dest, const void *src, size_t n);
+void		*ft_memset(void	*b, int c, size_t n);
+void		*ft_memmove(void *dst, const void *sc, size_t len);
+void		ft_putchar_fd(char c, int fd);
+void		ft_putstr_fd(char *str, int fd);
+void		ft_putendl_fd(char *str, int fd);
+void		ft_putnbr_fd(int nb, int fd);
+char		*ft_strchr(const char *str, int c);
+char		*ft_strdup(const char *src);
+char		*ft_strjoin(char const *s1, char const *s2);
+size_t		ft_strlcat(char *dest, const char *src, size_t size);
+size_t		ft_strlcpy(char *dest, const char *src, size_t size);
+size_t		ft_strlen(const char *str);
+int			ft_strncmp(const char *s1, const char *s2, size_t n);
+char		*ft_strnstr(const char *str, const char *find, size_t len);
+char		*ft_strrchr(const char *str, int c);
+char		*ft_strtrim(char const *s1, char const *set);
+char		*ft_substr(char const *str, unsigned int start, size_t len);
+char		*ft_strmapi(char const *str, char (*f)(unsigned int, char));
+void		ft_striteri(char *str, void (*f)(unsigned int, char*));
+char		**ft_split(char const *str, char c);
+int			ft_tolower(int c);
+int			ft_toupper(int c);
 
 #endif
